Move hp search to integrations
